+++ lib/searchkick/process_queue_job.rb
@@ -2,10 +2,11 @@ module Searchkick
   class ProcessQueueJob < ActiveJob::Base
     queue_as { Searchkick.queue_name }
 
-    def perform(class_name:, index_name: nil, inline: false)
+    def perform(class_name:, index_name: nil, inline: false, job_options: nil)
       model = Searchkick.load_model(class_name)
       index = model.searchkick_index(name: index_name)
       limit = model.searchkick_options[:batch_size] || 1000
+      job_options = (model.searchkick_options[:job_options] || {}).merge(job_options || {})
 
       loop do
         record_ids = index.reindex_queue.reserve(limit: limit)
@@ -20,7 +21,7 @@ module Searchkick
             # use new.perform to avoid excessive logging
             Searchkick::ProcessBatchJob.new.perform(**batch_options)
           else
-            Searchkick::ProcessBatchJob.perform_later(**batch_options)
+            Searchkick::ProcessBatchJob.set(job_options).perform_later(**batch_options)
           end
 
           # TODO when moving to reliable queuing, mark as complete